[增强功能] 回测引擎增加跨品种套利bar模式; 价差行情模块移动=> gateway.py; kline增加缠论支持

+++ vnpy/app/cta_strategy_pro/portfolio_testing.py
@@ -277,7 +277,7 @@ class PortfolioTestingEngine(BackTestingEngine):
                 bar.high_price = float(bar_data['high'])
                 bar.low_price = float(bar_data['low'])
                 bar.volume = int(bar_data['volume'])
-                bar.open_interest = int(bar_data.get('open_interest', 0))
+                bar.open_interest = float(bar_data.get('open_interest', 0))
                 bar.date = bar_datetime.strftime('%Y-%m-%d')
                 bar.time = bar_datetime.strftime('%H:%M:%S')
                 str_td = str(bar_data.get('trading_day', ''))
